admin record fix
