+++ argonath/views/admin.py
@@ -3,7 +3,7 @@
 from flask import url_for, redirect, g, render_template, Blueprint, flash, request, abort
 
 from argonath.utils import need_login, need_admin
-from argonath.models import User, Record, CIDR
+from argonath.models import User, CIDR
 
 bp = Blueprint('admin', __name__, url_prefix='/admin')
 
@@ -12,39 +12,6 @@ bp = Blueprint('admin', __name__, url_prefix='/admin')
 def index():
     users = User.list_users(g.start, g.limit)
     return render_template('admin.html', user=g.user, users=users)
-
-@bp.route('/create/', methods=['GET', 'POST'])
-@need_login
-def create_record():
-    if request.method == 'GET':
-        return render_template('admin_record.html')
-    name = request.form.get('name', type=str, default='').strip()
-    host_or_ip = request.form.get('host', type=str, default='').strip()
-    if len(name) < 3:
-        flash(u'域名长度必须大于3', 'error')
-        return redirect(url_for('admin.create_record'))
-    if name.startswith('.') or name.endswith('.'):
-        flash(u'域名不能以"."开始或者结束, 并不是dnspod啊 (￣▽￣")', 'error')
-        return redirect(url_for('admin.create_record'))
-
-    domain = name + '.hunantv.com'
-
-    r = Record.get_by_domain(domain)
-    if r:
-        flash(u'域名已经存在, 可以编辑', 'info')
-        return redirect(url_for('record.edit_record', record_id=r.id))
-    r = Record.create(g.user, name, domain, host_or_ip)
-    if not r:
-        flash(u'创建失败', 'error')
-        return redirect(url_for('admin.create_record'))
-    return redirect(url_for('record.get_record', record_id=r.id))
-
-@bp.route('/list/')
-@need_login
-def list_my_records():
-    records, total = g.user.list_records(g.start, g.limit)
-    admin_records = [r for r in records if r.domain.split('.')[1] == 'hunantv']
-    return render_template('admin_list.html', records=admin_records, total=total, endpoint='admin.list_my_records')
 
 @bp.route('/user_records/<username>/')
 @need_login
